+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - add required configuration, initial files and folders
 - add classes and methods to request tankerkönig api and store values in sqlite database
 - add internal api to retrieve the price data from database as JSON
+- add random delay in execution to fulfill the tankerkoenig.de terms of use
 
 ### Changed